Refactor link page tests for reuse
